GLFWwindow Hint Error
glfwWindowHint() was called before the creation of the glfw window;
Hence, the window was created with a version different from what our
code usses (most likely an older version that did not use vao).
Therefore, our code would display a black box for the texture instead of
the desired image (png).