Resource Manager Refactor:
Class ResourceManager was changed from being a static class to a on
class member variable.

Instead of being a global variable
available to all it will be used as a layer based resource manager. A
layer is a scene we wish to render such as: menu, ui overlay, game, etc.
Each layer or scene will have it own ResourceManager class member.